# backend/docs/openapi/reviews-api.yaml
openapi: 3.0.3
info:
  title: TeaSeat Reviews API
  version: 1.0.0
  description: Отзывы о товарах, оценки исполнения заказов и модерация.
servers:
  - url: /api
...
paths:
...
  /manager/reviews/{review}/hide:
    post:
      tags: [Manager moderation]
      operationId: hideReview
      x-laravel-route-name: manager.reviews.hide
      security: [{sanctum: []}]
      parameters: [{$ref: '#/components/parameters/ReviewId'}]
      requestBody: {$ref: '#/components/requestBodies/HideInput'}
      responses:
        '200': {description: Отзыв скрыт либо уже был скрыт}
  /manager/reviews/{review}/restore:
    post:
      tags: [Manager moderation]
      operationId: restoreReview
      x-laravel-route-name: manager.reviews.restore
      security: [{sanctum: []}]
      parameters: [{$ref: '#/components/parameters/ReviewId'}]
      requestBody: {$ref: '#/components/requestBodies/RestoreInput'}
      responses:
        '200': {description: Отзыв восстановлен либо уже был опубликован}
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
  parameters:
...
    ReviewId:
      in: path
      name: review
      required: true
      schema: {type: integer}
...
  requestBodies:
...
    HideInput:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [reason_code, comment]
            properties:
              reason_code:
                type: string
                enum: [spam, abuse, personal_data, off_topic, fraud, other]
              comment: {type: string, maxLength: 2000}
    RestoreInput:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [comment]
            properties:
              comment: {type: string, maxLength: 2000}
...